# internal/app/docs/swagger.yaml
basePath: /
definitions:
  github_com_eren_dev_go_server_internal_shared_pagination.PaginationInfo:
    properties:
      limit:
        description: Límite de registros por página
        example: 10
        type: integer
      skip:
        description: Registros saltados
        example: 0
        type: integer
      total:
        description: Total de registros
        example: 100
        type: integer
      total_pages:
        description: Total de páginas
        example: 10
        type: integer
    type: object
  github_com_eren_dev_go_server_internal_shared_validation.FieldError:
    properties:
      field:
        description: Nombre del campo con error
        example: email
        type: string
      message:
        description: Mensaje de error
        example: must be a valid email
        type: string
    type: object
  github_com_eren_dev_go_server_internal_shared_validation.ValidationError:
    properties:
      errors:
        description: Lista de errores por campo
        items:
          $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.FieldError'
        type: array
    type: object
...
  internal_modules_patients.CreatePatientDTO:
    properties:
      avatar_url:
        type: string
      birth_date:
        type: string
      breed:
        example: Labrador
        type: string
      color:
        example: Dorado
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/internal_modules_patients.Gender'
        enum:
        - male
        - female
        - unknown
      microchip:
        example: "985112000123456"
        type: string
      name:
        example: Max
        type: string
      notes:
        type: string
      owner_id:
        type: string
      species_id:
        type: string
      sterilized:
        type: boolean
      weight:
        example: 12.5
        type: number
    required:
    - gender
    - name
    - owner_id
    - species_id
    - weight
    type: object
  internal_modules_patients.Gender:
    enum:
    - male
    - female
    - unknown
    type: string
    x-enum-varnames:
    - GenderMale
    - GenderFemale
    - GenderUnknown
  internal_modules_patients.PaginatedPatientsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_modules_patients.PatientResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_pagination.PaginationInfo'
    type: object
  internal_modules_patients.PatientResponse:
    properties:
      active:
        type: boolean
      avatar_url:
        type: string
      birth_date:
        type: string
      breed:
        type: string
      color:
        type: string
      created_at:
        type: string
      gender:
        $ref: '#/definitions/internal_modules_patients.Gender'
      id:
        type: string
      microchip:
        type: string
      name:
        type: string
      notes:
        type: string
      owner_id:
        type: string
      species_id:
        type: string
      sterilized:
        type: boolean
      tenant_id:
        type: string
      updated_at:
        type: string
      weight:
        type: number
    type: object
  internal_modules_patients.ResolveSpeciesDTO:
    properties:
      name:
        example: Canino
        type: string
    required:
    - name
    type: object
  internal_modules_patients.SpeciesResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      normalized_name:
        type: string
      tenant_id:
        type: string
    type: object
  internal_modules_patients.UpdatePatientDTO:
    properties:
      active:
        type: boolean
      avatar_url:
        type: string
      birth_date:
        type: string
      breed:
        type: string
      color:
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/internal_modules_patients.Gender'
        enum:
        - male
        - female
        - unknown
      microchip:
        type: string
      name:
        type: string
      notes:
        type: string
      species_id:
        type: string
      sterilized:
        type: boolean
      weight:
        type: number
    type: object
  internal_modules_payments.CreatePaymentDTO:
    properties:
      amount:
        example: 4900
        minimum: 1
        type: integer
      concept:
        example: Pago mensual - Plan Pro
        type: string
      currency:
        example: USD
        type: string
      external_transaction_id:
        example: wompi_tx_123456
        type: string
      failure_reason:
        example: Fondos insuficientes
        type: string
      metadata:
        additionalProperties: true
        type: object
      payment_method:
        example: wompi
        type: string
      period_end:
        example: "2024-02-01T00:00:00Z"
        type: string
      period_start:
        example: "2024-01-01T00:00:00Z"
        type: string
      processed_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      status:
        allOf:
        - $ref: '#/definitions/internal_modules_payments.PaymentStatus'
        example: completed
      tenant_id:
        example: 507f1f77bcf86cd799439011
        type: string
    required:
    - amount
    - concept
    - currency
    - payment_method
    - status
    - tenant_id
    type: object
  internal_modules_payments.PaymentResponse:
    properties:
      amount:
        example: 4900
        type: integer
      concept:
        example: Pago mensual - Plan Pro
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      currency:
        example: USD
        type: string
      external_transaction_id:
        example: wompi_tx_123456
        type: string
      failure_reason:
        example: Fondos insuficientes
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      metadata:
        additionalProperties: true
        type: object
      payment_method:
        example: wompi
        type: string
      period_end:
        example: "2024-02-01T00:00:00Z"
        type: string
      period_start:
        example: "2024-01-01T00:00:00Z"
        type: string
      processed_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      status:
        allOf:
        - $ref: '#/definitions/internal_modules_payments.PaymentStatus'
        example: completed
      tenant_id:
        example: 507f1f77bcf86cd799439011
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  internal_modules_payments.PaymentStatus:
    enum:
    - pending
    - completed
    - failed
    - refunded
    type: string
    x-enum-varnames:
    - PaymentPending
    - PaymentCompleted
    - PaymentFailed
    - PaymentRefunded
...
host: localhost:8080
info:
  contact:
    email: support@vetsify.com
    name: Vetsify Support
  description: |-
    API para plataforma veterinaria multi-tenant. Soporta panel de administración (staff) y app móvil (owners).

    ## Autenticación
    Todas las rutas protegidas requieren el header `Authorization: Bearer <token>`.
    - **Staff**: Obtén tokens via `POST /api/auth/login`
    - **Owners (mobile)**: Obtén tokens via `POST /mobile/auth/login`

    ## Multi-tenancy
    Las rutas que operan sobre datos de un tenant específico (patients, species, etc.) requieren el header `X-Tenant-ID` con el ObjectID del tenant.
    Este header debe enviarse en **todas** las peticiones a rutas marcadas con el parámetro `X-Tenant-ID`.
    El frontend puede configurarlo una vez en el interceptor HTTP (axios/fetch).

    ## Roles (RBAC)
    Las rutas del panel admin (`/api/*`) están protegidas por RBAC.
    Roles disponibles: `admin`, `veterinarian`, `receptionist`, `assistant`, `accountant`.

    ## Paginación
    Las rutas que retornan listas soportan `skip` y `limit` como query params.
    Respuesta incluye `pagination: { skip, limit, total, total_pages }`.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Vetsify API
  version: "1.0"
paths:
...
  /api/patients:
    get:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Skip
        in: query
        name: skip
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_patients.PaginatedPatientsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List patients
      tags:
      - patients
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Patient data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_patients.CreatePatientDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_patients.PatientResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create patient
      tags:
      - patients
  /api/patients/{id}:
    delete:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Patient ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete patient
      tags:
      - patients
    get:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Patient ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_patients.PatientResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get patient by ID
      tags:
      - patients
    patch:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Patient ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_patients.UpdatePatientDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_patients.PatientResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update patient
      tags:
      - patients
  /api/payments:
    post:
      consumes:
      - application/json
      description: Registra un nuevo pago en el sistema
      parameters:
      - description: Datos del pago
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/internal_modules_payments.CreatePaymentDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_modules_payments.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear pago
      tags:
      - payments
  /api/payments/{id}:
    get:
      description: Obtiene un pago por ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_payments.PaymentResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener pago
      tags:
      - payments
...
  /api/species:
    get:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_modules_patients.SpeciesResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List species
      tags:
      - species
    post:
      consumes:
      - application/json
      description: Finds an exact/similar match or creates a new species. Returns
        409 with suggestions if ambiguous.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Species name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_patients.ResolveSpeciesDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_patients.SpeciesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Resolve or create species
      tags:
      - species
...
  /api/tenants/{tenant_id}/payments:
    get:
      description: Obtiene el historial de pagos de un tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - default: 50
        description: Límite de resultados
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_modules_payments.PaymentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Historial de pagos de un tenant
      tags:
      - payments
...
  /api/webhooks/{provider}:
    post:
      consumes:
      - application/json
      description: Endpoint para recibir webhooks de proveedores de pago
      parameters:
      - description: Provider name
        enum:
        - wompi
        - stripe
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Procesar webhook de pago
      tags:
      - webhooks
...
  /mobile/patients:
    get:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Skip
        in: query
        name: skip
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_patients.PaginatedPatientsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List my patients
      tags:
      - mobile/patients
  /mobile/patients/{id}:
    get:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Patient ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_patients.PatientResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get my patient by ID
      tags:
      - mobile/patients
securityDefinitions:
  Bearer:
    description: 'Ingresa el token con el prefijo Bearer: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"